Lesson11

+++ app/src/main/java/ru/gb/notesapp/NotesFragment.java
@@ -1,12 +1,14 @@
 package ru.gb.notesapp;
 
 import android.app.Activity;
+import android.content.DialogInterface;
 import android.content.res.Configuration;
 import android.graphics.Color;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
@@ -49,7 +51,6 @@ public class NotesFragment extends Fragment {
     private RecyclerView recyclerView;
 
 
-
     public NotesFragment() {
 
     }
@@ -61,7 +62,7 @@ public class NotesFragment extends Fragment {
 
         View view = inflater.inflate(R.layout.fragment_recycleview, container, false);
 
-            initList(view, savedInstanceState);
+        initList(view, savedInstanceState);
         return view;
 
 
@@ -77,7 +78,6 @@ public class NotesFragment extends Fragment {
         fragmentTransaction.commit();
 
 
-
         isLand = getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE;
         if (isLand) {
             showNotesContentLand(DEFAULT_INDEX);
@@ -89,11 +89,10 @@ public class NotesFragment extends Fragment {
     private void initList(View view, Bundle savedInstanceState) {
 
 
-
         recyclerView = view.findViewById(R.id.recycler_view_lines);
 //        CardsSourceImpl cardsSource = new CardsSourceImpl(getResources());
 //        CardsSourceImpl cardsSource = CardsSourceImpl.getInstance(getResources());
-        CardSource cardsSource =  CardsSourceFirebaseImpl.getInstance();
+        CardSource cardsSource = CardsSourceFirebaseImpl.getInstance();
 
         cardsSource.init(cardsData -> adapter.notifyDataSetChanged());
 //        data = cardsSource.getData() ;
@@ -107,35 +106,36 @@ public class NotesFragment extends Fragment {
 
         setHasOptionsMenu(true);
 
-        adapter = new ItemAdapter(data,this);
+        adapter = new ItemAdapter(data, this);
         LinearLayoutManager layoutManager = new LinearLayoutManager(getContext());
 
         recyclerView.setLayoutManager(layoutManager);
         recyclerView.setAdapter(adapter);
 
-        DividerItemDecoration itemDecoration = new DividerItemDecoration(getContext(),LinearLayoutManager.VERTICAL);
-        itemDecoration.setDrawable(getResources().getDrawable(R.drawable.separator,null));
+        DividerItemDecoration itemDecoration = new DividerItemDecoration(getContext(), LinearLayoutManager.VERTICAL);
+        itemDecoration.setDrawable(getResources().getDrawable(R.drawable.separator, null));
         recyclerView.addItemDecoration(itemDecoration);
 
         adapter.setListener(position -> {
             showNotesContent(position);
 //                updateText(position);
         });
-        adapter.setLongClickListener(position -> {});
+        adapter.setLongClickListener(position -> {
+        });
 
 
     }
 
     @Override
-    public void onCreateOptionsMenu(@NonNull  Menu menu, @NonNull MenuInflater inflater) {
+    public void onCreateOptionsMenu(@NonNull Menu menu, @NonNull MenuInflater inflater) {
         inflater.inflate(R.menu.menu, menu);
     }
 
     @Override
-    public boolean onOptionsItemSelected(@NonNull  MenuItem item) {
-        switch (item.getItemId()){
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        switch (item.getItemId()) {
             case R.id.add_note:
-                CardData cardData = new CardData("Simple notes","Add your text");
+                CardData cardData = new CardData("Simple notes", "Add your text");
 //                CardData cardContentData = new CardData("Add Text");
 
                 cardData.setId(UUID.randomUUID().toString());
@@ -144,8 +144,8 @@ public class NotesFragment extends Fragment {
 //                data.addCardContentData(cardContentData);
 //                data.addCardData(new CardData("Simple notes"));
 //                data.addCardContentData(new CardData("Simple notes"));
-                adapter.notifyItemInserted(data.size()-1);
-                recyclerView.scrollToPosition(data.size()-1);
+                adapter.notifyItemInserted(data.size() - 1);
+                recyclerView.scrollToPosition(data.size() - 1);
                 return true;
             case R.id.delete_all:
                 data.clearCardData();
@@ -157,7 +157,7 @@ public class NotesFragment extends Fragment {
     }
 
     @Override
-    public void onCreateContextMenu(@NonNull  ContextMenu menu, @NonNull  View v, @Nullable ContextMenu.ContextMenuInfo menuInfo) {
+    public void onCreateContextMenu(@NonNull ContextMenu menu, @NonNull View v, @Nullable ContextMenu.ContextMenuInfo menuInfo) {
         super.onCreateContextMenu(menu, v, menuInfo);
         MenuInflater inflater = requireActivity().getMenuInflater();
         inflater.inflate(R.menu.context_menu, menu);
@@ -166,14 +166,39 @@ public class NotesFragment extends Fragment {
 
     @Override
     public boolean onContextItemSelected(@NonNull MenuItem item) {
+
+
         if (item.getItemId() == R.id.item1_context_menu) {
-            currentPosition = adapter.getCurrentPosition();
-            data.deleteCardData(currentPosition);
-            adapter.notifyItemRemoved(currentPosition);
-            Toast.makeText(getContext(), "Note has been deleted" + currentPosition, Toast.LENGTH_SHORT).show();
+
+           AlertDialog.Builder builder = new AlertDialog.Builder(requireContext());
+           builder.setMessage(R.string.message)
+                   .setTitle(R.string.tittle)
+                   .setCancelable(false)
+                   .setPositiveButton(R.string.ok, (dialog, which) -> {
+                       currentPosition = adapter.getCurrentPosition();
+                       data.deleteCardData(currentPosition);
+                       adapter.notifyItemRemoved(currentPosition);
+                       Toast.makeText(getContext(),R.string.message_note_deleted, Toast.LENGTH_SHORT).show();
+
+                   })
+                   .setNegativeButton(R.string.cancel, (dialog, which) -> {
+
+                   })
+                   .show();
+
             return true;
+
         }
         return super.onContextItemSelected(item);
+
+//        if (item.getItemId() == R.id.item1_context_menu) {
+//            currentPosition = adapter.getCurrentPosition();
+//            data.deleteCardData(currentPosition);
+//            adapter.notifyItemRemoved(currentPosition);
+//            Toast.makeText(getContext(), "Note has been deleted" + currentPosition, Toast.LENGTH_SHORT).show();
+//            return true;
+//        }
+//        return super.onContextItemSelected(item);
     }
 
 //    public void initPopupMenu() {
@@ -192,9 +217,6 @@ public class NotesFragment extends Fragment {
 //    }
 
 
-
-
-
     private void showNotesContent(int finalIndex) {
         if (isLand) {
             showNotesContentLand(finalIndex);
@@ -209,7 +231,7 @@ public class NotesFragment extends Fragment {
         NotesContentFragment fragment = NotesContentFragment.newInstance(index);
         requireActivity().getSupportFragmentManager()
                 .beginTransaction()
-                .add(R.id.notes_fragments_container,fragment)
+                .add(R.id.notes_fragments_container, fragment)
                 .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
                 .addToBackStack(null)
                 .commit();
